[UPDATE] MapService 통합
