+++ src/main/java/UMC_8th/With_Run/map/repository/MapRepository.java
@@ -1,4 +1,4 @@
 package UMC_8th.With_Run.map.repository;
 
-public class MapRepository {
+public class MapRepository{
 }
