+++ src/main/java/UMC_8th/With_Run/map/repository/PinRepository.java
@@ -2,8 +2,6 @@ package UMC_8th.With_Run.map.repository;
 
 import UMC_8th.With_Run.map.entity.Pin;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface PinRepository extends JpaRepository<Pin, Long> {
 }
