+++ src/main/java/UMC_8th/With_Run/map/service/MapServiceImpl.java
@@ -5,50 +5,30 @@ import UMC_8th.With_Run.common.exception.handler.MapHandler;
 import UMC_8th.With_Run.map.dto.MapRequestDTO;
 import UMC_8th.With_Run.map.dto.MapResponseDTO;
 import UMC_8th.With_Run.map.entity.PetFacility;
+import UMC_8th.With_Run.map.entity.Pin;
 import UMC_8th.With_Run.map.repository.PetFacilityRepository;
+import UMC_8th.With_Run.map.repository.PinRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class MapSearchServiceImpl implements MapSearchService {
+public class MapServiceImpl implements MapService {
 
     private final PetFacilityRepository petFacilityRepository;
+    private final PinRepository pinRepository;
     private final CourseService courseService;
 
     @Override
-    public List<PetFacility> searchPlacesByCategory(String category) {
-        return petFacilityRepository.findByCategoryContainingIgnoreCase(category);
-    }
-
-    @Override
-    public Page<MapResponseDTO.PetFacilityResponseDto> searchPlacesByCategory(String category, int page, int size) {
-        validatePagingParameters(page, size);
-        
-        Pageable pageable = PageRequest.of(page, size);
-        Page<PetFacility> facilities = petFacilityRepository.findByCategoryContainingIgnoreCase(category, pageable);
-
-        return facilities.map(facility -> MapResponseDTO.PetFacilityResponseDto.builder()
-                .id(facility.getId())
-                .name(facility.getName())
-                .category(facility.getCategory())
-                .longitude(facility.getLongitude())
-                .latitude(facility.getLatitude())
-                .address(facility.getAddress())
-                .closed_day(facility.getClosed_day())
-                .running_time(facility.getRunning_time())
-                .has_parking(facility.getHas_parking())
-                .build());
-    }
-
-    @Override
-    public MapResponseDTO.PetFacilityPageResponseDto searchPlacesByCategorySimple(String category, int page, int size) {
+    public MapResponseDTO.PetFacilityPageResponseDto getPetFacilityByCategory(String category, int page, int size) {
         validatePagingParameters(page, size);
         
         Pageable pageable = PageRequest.of(page, size);
@@ -117,5 +97,70 @@ public class MapSearchServiceImpl implements MapSearchService {
     @Override
     public Long createCourse(MapRequestDTO.CourseCreateRequestDto requestDto) {
         return courseService.createCourse(requestDto.getUserId(), requestDto);
+    }
+
+    @Override
+    @Transactional
+    public Long createPin(MapRequestDTO.PinRequestDto requestDto) {
+        Pin pin = Pin.builder()
+                .name(requestDto.getName())
+                .detail(requestDto.getDetail())
+                .color(requestDto.getColor())
+                .latitude(requestDto.getLatitude())
+                .longitude(requestDto.getLongitude())
+                .createdAt(LocalDateTime.now())
+                .build();
+
+        pinRepository.save(pin);
+
+        return pin.getId();
+    }
+
+    @Override
+    @Transactional
+    public Long updatePin(Long pinId, MapRequestDTO.PinRequestDto requestDto) {
+        Pin pin = pinRepository.findById(pinId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 핀 없음"));
+        pin.setName(requestDto.getName());
+        pin.setDetail(requestDto.getDetail());
+        pin.setColor(requestDto.getColor());
+        pin.setLatitude(requestDto.getLatitude());
+        pin.setLongitude(requestDto.getLongitude());
+        pin.setUpdatedAt(LocalDateTime.now());
+
+        pinRepository.save(pin);
+        return pin.getId();
+    }
+
+    @Override
+    @Transactional
+    public Long deletePin(Long pinId) {
+        Pin pin = pinRepository.findById(pinId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 핀 없음"));
+
+        pinRepository.deleteById(pinId); // 물리적 삭제
+
+        return pin.getId(); // 삭제된 핀의 ID를 반환
+    }
+
+
+    @Override
+    @Transactional(readOnly = true) // 데이터 조회만 하므로 readOnly=true 설정
+    public MapResponseDTO.GetPinDto getPinById(Long pinId) {
+        Pin pin = pinRepository.findById(pinId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 핀 없음"));
+        return fromEntity(pin);
+    }
+
+    public static MapResponseDTO.GetPinDto fromEntity(Pin pin) {
+        return MapResponseDTO.GetPinDto.builder()
+                .pinId(pin.getId())
+                .courseId(pin.getCourseId())
+                .name(pin.getName())
+                .detail(pin.getDetail())
+                .color(pin.getColor())
+                .latitude(pin.getLatitude())
+                .longitude(pin.getLongitude())
+                .build();
     }
 }
