+++ src/main/resources/application.yml
@@ -21,7 +21,7 @@ spring:
         format_sql: true
         use_sql_comments: true
         hbm2ddl:
-          auto: none
+          auto: update
         default_batch_fetch_size: 1000
 
 jwt:
